+++ Dfe.Academies.Academisation.Service/Commands/ConversionProject/SetSchoolOverviewCommand.cs
@@ -11,7 +11,7 @@ namespace Dfe.Academies.Academisation.Service.Commands.ConversionProject
 			string viabilityIssues,
 			string partOfPfiScheme,
 			string financialDeficit,
-			int? numberOfPlacesFundedFor,
+			decimal? numberOfPlacesFundedFor,
 			string pfiSchemeDetails,
 			decimal? distanceFromSchoolToTrustHeadquarters,
 			string distanceFromSchoolToTrustHeadquartersAdditionalInformation,
@@ -34,7 +34,7 @@ namespace Dfe.Academies.Academisation.Service.Commands.ConversionProject
 		public string ViabilityIssues { get; set; }
 		public string PartOfPfiScheme { get; set; }
 		public string FinancialDeficit { get; set; }
-		public int? NumberOfPlacesFundedFor { get; set; }
+		public decimal? NumberOfPlacesFundedFor { get; set; }
 		public string PfiSchemeDetails { get; set; }
 		public decimal? DistanceFromSchoolToTrustHeadquarters { get; set; }
 		public string DistanceFromSchoolToTrustHeadquartersAdditionalInformation { get; set; }
